Renamed ConfigParser to match other modules.
